requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-graphql-jwt==0.2.0
djangorestframework==3.9.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
Mako==1.0.7
MarkupSafe==1.0
oauth==1.0.1
olefile==0.45.1
pbr==5.1.1
pexpect==4.2.1
Pillow==5.1.0
promise==2.2.1
psutil==5.4.2
pycrypto==2.6.1
PyMySQL==0.9.3
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
Rx==1.6.1
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.22
virtualenv==16.2.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wadllib==1.3.2
zope.interface==4.3.2
